using Script and command as a node step plugin cleaning
