+++ plugins/script-node-step-plugin/src/main/groovy/org/rundeck/plugin/scriptnodestep/CommandNodeStepPlugin.groovy
@@ -4,26 +4,18 @@ import com.dtolabs.rundeck.core.common.INodeEntry
 import com.dtolabs.rundeck.core.dispatcher.DataContextUtils
 import com.dtolabs.rundeck.core.execution.ExecArgList
 import com.dtolabs.rundeck.core.execution.ExecCommand
-import com.dtolabs.rundeck.core.execution.ExecutionContext
-import com.dtolabs.rundeck.core.execution.ExecutionServiceImpl
-import com.dtolabs.rundeck.core.execution.proxy.ProxyRunnerPlugin
 import com.dtolabs.rundeck.core.execution.service.NodeExecutorResult
 import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepException
-import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepResult
 import com.dtolabs.rundeck.core.plugins.Plugin
-import com.dtolabs.rundeck.core.utils.OptsUtil
-import com.dtolabs.rundeck.plugins.ExecutionEnvironmentConstants
 import com.dtolabs.rundeck.plugins.ServiceNameConstants
 import com.dtolabs.rundeck.plugins.descriptions.PluginDescription
-import com.dtolabs.rundeck.plugins.descriptions.PluginMetadata
 import com.dtolabs.rundeck.plugins.descriptions.PluginProperty
 import com.dtolabs.rundeck.plugins.step.NodeStepPlugin
 import com.dtolabs.rundeck.plugins.step.PluginStepContext
 
 @Plugin(service = ServiceNameConstants.WorkflowNodeStep, name = EXEC_COMMAND_TYPE)
 @PluginDescription(title = "Command", description = "Run a command on the remote node", isHighlighted = true, order = 0)
-@PluginMetadata(key = ExecutionEnvironmentConstants.ENVIRONMENT_TYPE_KEY, value = ExecutionEnvironmentConstants.LOCAL_RUNNER)
-class CommandNodeStepPlugin implements NodeStepPlugin, ExecCommand, ProxyRunnerPlugin {
+class CommandNodeStepPlugin extends ScriptProxyRunner implements NodeStepPlugin, ExecCommand {
 
     @PluginProperty(title = "Command",
             description = "",
@@ -47,13 +39,4 @@ class CommandNodeStepPlugin implements NodeStepPlugin, ExecCommand, ProxyRunnerP
         }
     }
 
-    @Override
-    Map<String, String> getRuntimeProperties(ExecutionContext context) {
-        return context.getFramework().getFrameworkProjectMgr().loadProjectConfig(context.frameworkProject).getProjectProperties()
-    }
-
-    @Override
-    Map<String, String> getRuntimeFrameworkProperties(ExecutionContext context){
-        return context.getIFramework().getPropertyLookup().getPropertiesMap()
-    }
 }
