+++ plugins/script-node-step-plugin/src/main/groovy/org/rundeck/plugin/scriptnodestep/ScriptFileNodeStepExecutor.groovy
@@ -4,17 +4,18 @@ import com.dtolabs.rundeck.core.common.INodeEntry;
 import com.dtolabs.rundeck.core.data.SharedDataContextUtils;
 import com.dtolabs.rundeck.core.dispatcher.ContextView;
 import com.dtolabs.rundeck.core.dispatcher.DataContextUtils
-import com.dtolabs.rundeck.core.execution.ExecutionContextImpl
 import com.dtolabs.rundeck.core.execution.ExecutionService
-import com.dtolabs.rundeck.core.execution.ExecutionServiceImpl
-import com.dtolabs.rundeck.core.execution.service.NodeExecutorResult
 import com.dtolabs.rundeck.core.execution.workflow.StepExecutionContext
-import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepException;
+import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepException
+import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepResult;
 import com.dtolabs.rundeck.core.execution.workflow.steps.node.impl.DefaultScriptFileNodeStepUtils;
 import com.dtolabs.rundeck.core.utils.OptsUtil;
 import com.dtolabs.rundeck.plugins.step.PluginStepContext
+import groovy.transform.CompileDynamic
+import groovy.transform.CompileStatic
 
-public class ScriptFileNodeStepExecutor {
+@CompileStatic
+class ScriptFileNodeStepExecutor {
     private final String scriptInterpreter;
     private final Boolean interpreterArgsQuoted;
     private final String fileExtension;
@@ -25,7 +26,7 @@ public class ScriptFileNodeStepExecutor {
 
     private final DefaultScriptFileNodeStepUtils scriptUtils = new DefaultScriptFileNodeStepUtils();
 
-public ScriptFileNodeStepExecutor(
+    ScriptFileNodeStepExecutor(
             String scriptInterpreter,
             Boolean interpreterArgsQuoted,
             String fileExtension,
@@ -43,7 +44,8 @@ public ScriptFileNodeStepExecutor(
         this.expandTokenInScriptFile = expandTokenInScriptFile;
     }
 
-    public void executeScriptFile(PluginStepContext context, Map<String, Object> configuration, INodeEntry entry) {
+    @CompileDynamic
+    void executeScriptFile(PluginStepContext context, Map<String, Object> configuration, INodeEntry entry) {
         boolean expandTokens = true;
         if (context.getFramework().hasProperty("execution.script.tokenexpansion.enabled")) {
             expandTokens = "true".equals(context.getFramework().getProperty("execution.script.tokenexpansion.enabled"));
@@ -81,7 +83,7 @@ public ScriptFileNodeStepExecutor(
 
         boolean argsQuoted = interpreterArgsQuoted != null ? interpreterArgsQuoted : false;
 
-        NodeExecutorResult nodeExecutorResult = scriptUtils.executeScriptFile(
+        NodeStepResult nodeExecutorResult = scriptUtils.executeScriptFile(
                 stepExecutionContext,
                 entry,
                 this.adhocLocalString,
@@ -95,7 +97,7 @@ public ScriptFileNodeStepExecutor(
                 expandTokens
         );
 
-        if(nodeExecutorResult.resultCode != 0){
+        if(!nodeExecutorResult.isSuccess()){
             throw new NodeStepException( nodeExecutorResult.failureMessage, nodeExecutorResult.failureReason, entry.getNodename())
         }
     }
