+++ plugins/script-node-step-plugin/src/main/groovy/org/rundeck/plugin/scriptnodestep/ScriptFileNodeStepPlugin.groovy
@@ -17,10 +17,8 @@
 package org.rundeck.plugin.scriptnodestep
 
 import com.dtolabs.rundeck.core.common.INodeEntry
-import com.dtolabs.rundeck.core.execution.ExecutionContext
 import com.dtolabs.rundeck.core.execution.ScriptFileCommand
 import com.dtolabs.rundeck.core.execution.proxy.ProxyRunnerPlugin
-import com.dtolabs.rundeck.core.execution.proxy.ProxySecretBundleCreator
 import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepException
 import com.dtolabs.rundeck.core.plugins.Plugin
 import com.dtolabs.rundeck.plugins.ServiceNameConstants
@@ -30,9 +28,9 @@ import com.dtolabs.rundeck.plugins.step.NodeStepPlugin
 import com.dtolabs.rundeck.plugins.step.PluginStepContext
 
 
-@Plugin(service = ServiceNameConstants.WorkflowNodeStep, name = ScriptFileNodeStepPlugin.SCRIPT_FILE_COMMAND_TYPE)
+@Plugin(service = ServiceNameConstants.WorkflowNodeStep, name = SCRIPT_FILE_COMMAND_TYPE)
 @PluginDescription(title = "Script file or URL", description = "Verify and validate design", isHighlighted = true, order = 2)
-class ScriptFileNodeStepPlugin implements NodeStepPlugin, ScriptFileCommand, ProxyRunnerPlugin {
+class ScriptFileNodeStepPlugin extends ScriptProxyRunner implements NodeStepPlugin, ScriptFileCommand, ProxyRunnerPlugin {
 
     @PluginProperty(title = "File Path or URL",
             description = "Path",
@@ -110,44 +108,5 @@ class ScriptFileNodeStepPlugin implements NodeStepPlugin, ScriptFileCommand, Pro
     Boolean getAdhocExecution() {
         return null
     }
-
-    @Override
-    Map<String, String> getRuntimeProperties(ExecutionContext context) {
-        return context.getFramework().getFrameworkProjectMgr().loadProjectConfig(context.frameworkProject).getProjectProperties()
-    }
-
-    @Override
-    Map<String, String> getRuntimeFrameworkProperties(ExecutionContext context){
-        return context.getIFramework().getPropertyLookup().getPropertiesMap()
-    }
-
-    @Override
-    //get shared secrets from the original node-executor plugin
-    List<String> listSecretsPathWorkflowNodeStep(ExecutionContext context, INodeEntry node, Map<String, Object> configuration) {
-
-        def executionService = context.getFramework().getNodeExecutorService()
-        //get original node executor from node or project
-        String orig = executionService.getDefaultProviderNameForNodeAndProject(node, context.getFrameworkProject())
-        if (null != node.getAttributes() && null != node.getAttributes().get(executionService.getServiceProviderNodeAttributeForNode(
-                node))) {
-            orig = node.getAttributes().get(executionService.getServiceProviderNodeAttributeForNode(node));
-        }
-        //get provider
-        def provider = executionService.providerOfType(orig)
-
-        def list = new ArrayList<String>()
-        if(provider instanceof ProxyRunnerPlugin){
-            //get list of secrets from original node-executor plugin
-            list = provider.listSecretsPath(context, node)
-        }
-
-        if(provider instanceof ProxySecretBundleCreator){
-            //get list of secrets from original node-executor plugin
-            list = provider.listSecretsPath(context, node)
-        }
-
-        return list
-    }
-
 
 }
